AUTONOMOUS ONLY PlayerMovement -- not working currently
This is a new version of PlayerMovement that only traverses the maze autonomously. It is currently not functioning correctly as I believe in the DetermineDirection, the case that is supposed to handle when the robot starts moving is being called at all times, making the robot move towards the farthest wall away, eventually stopping at the midpoint of a corridor.